migrate code to RoutedTabs
